+++ test/kitchensink/babel.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  presets: [
-    require.resolve('../../core/poi/babel'),
-    'power-assert'
-  ]
-}