+++ core/poi/CHANGELOG.md
@@ -3,6 +3,83 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.2.13](https://github.com/egoist/poi/compare/poi@12.2.12...poi@12.2.13) (2019-01-02)
+
+### Features
+
+- allow config file to export a function ([8e82f66](https://github.com/egoist/poi/commit/8e82f66))
+
+## [12.2.12](https://github.com/egoist/poi/compare/poi@12.2.11...poi@12.2.12) (2018-12-31)
+
+**Note:** Version bump only for package poi
+
+## [12.2.11](https://github.com/egoist/poi/compare/poi@12.2.10...poi@12.2.11) (2018-12-31)
+
+### Bug Fixes
+
+- use recommended bucklescript workflow ([2231c5c](https://github.com/egoist/poi/commit/2231c5c))
+
+## [12.2.10](https://github.com/egoist/poi/compare/poi@12.2.9...poi@12.2.10) (2018-12-31)
+
+### Bug Fixes
+
+- let webpack watch all reason files ([e5c7a05](https://github.com/egoist/poi/commit/e5c7a05))
+
+## [12.2.9](https://github.com/egoist/poi/compare/poi@12.2.8...poi@12.2.9) (2018-12-30)
+
+### Bug Fixes
+
+- cache support for vue templates ([aa0a430](https://github.com/egoist/poi/commit/aa0a430))
+
+## [12.2.8](https://github.com/egoist/poi/compare/poi@12.2.7...poi@12.2.8) (2018-12-28)
+
+### Features
+
+- use source field from package.json as default app entry ([af61639](https://github.com/egoist/poi/commit/af61639))
+
+## [12.2.7](https://github.com/egoist/poi/compare/poi@12.2.6...poi@12.2.7) (2018-12-25)
+
+### Bug Fixes
+
+- **cli:** help message ([648727e](https://github.com/egoist/poi/commit/648727e))
+
+### Features
+
+- add a hook to allow modifing devServer config ([df939e6](https://github.com/egoist/poi/commit/df939e6))
+
+## [12.2.6](https://github.com/egoist/poi/compare/poi@12.2.5...poi@12.2.6) (2018-12-23)
+
+### Bug Fixes
+
+- **cli:** cache is enabled by default ([b2171d5](https://github.com/egoist/poi/commit/b2171d5))
+- properly handle --no-config flag ([bb94787](https://github.com/egoist/poi/commit/bb94787))
+
+## [12.2.5](https://github.com/egoist/poi/compare/poi@12.2.4...poi@12.2.5) (2018-12-22)
+
+### Bug Fixes
+
+- only change devtoolModuleFilenameTemplate under --serve flag ([fe5b708](https://github.com/egoist/poi/commit/fe5b708))
+- resolve modules in local development ([369a352](https://github.com/egoist/poi/commit/369a352))
+
+### Features
+
+- add a new plugin option `when` ([17ff228](https://github.com/egoist/poi/commit/17ff228))
+- support all webpack targets ([0d51120](https://github.com/egoist/poi/commit/0d51120))
+
+## [12.2.4](https://github.com/egoist/poi/compare/poi@12.2.3...poi@12.2.4) (2018-12-21)
+
+### Bug Fixes
+
+- react-error-overlay doesn't work with eval source map ([ddc8213](https://github.com/egoist/poi/commit/ddc8213))
+- use original file path in sourcemap in dev mode ([a2dfc5f](https://github.com/egoist/poi/commit/a2dfc5f))
+
+## [12.2.3](https://github.com/egoist/poi/compare/poi@12.2.2...poi@12.2.3) (2018-12-21)
+
+### Bug Fixes
+
+- better naming for server hooks ([b9aab0b](https://github.com/egoist/poi/commit/b9aab0b))
+- **cli:** only open browser on the first successful build ([13b1ac8](https://github.com/egoist/poi/commit/13b1ac8))
+
 ## [12.2.2](https://github.com/egoist/poi/compare/poi@12.2.1...poi@12.2.2) (2018-12-19)
 
 ### Bug Fixes